+++ src/game/world/resources/copper.rs
@@ -16,16 +16,16 @@ impl WorldResource for Copper {
     "Copper"
   }
 
-  fn terrain_tile(&self, world: &World, position: [i32; 2]) -> TerrainTile {
+  fn terrain_tile(&self, world: &World, position: [i64; 2]) -> TerrainTile {
     let value = self.get_complex_tile_value(world, position, 1000..6000);
     TerrainTile::Complex(ComplexTerrainTile::Copper(value))
   }
 
-  fn get_tile(&self, world: &World, position: [i32; 2], base_terrain_mod: f64) -> bool {
+  fn get_tile(&self, world: &World, position: [i64; 2], base_terrain_mod: f64) -> bool {
     self.get_value(world, position) - base_terrain_mod > 0.8
   }
 
-  fn get_value(&self, world: &World, [x, y]: [i32; 2]) -> f64 {
+  fn get_value(&self, world: &World, [x, y]: [i64; 2]) -> f64 {
     const NOISE_SCALE: f64 = 0.04;
     world.noise_gen.get([x as f64 * NOISE_SCALE, y as f64 * NOISE_SCALE])
   }
