+++ src/game/world/resources/water.rs
@@ -16,15 +16,15 @@ impl WorldResource for Water {
     "Water"
   }
 
-  fn terrain_tile(&self, _: &World, _: [i32; 2]) -> TerrainTile {
+  fn terrain_tile(&self, _: &World, _: [i64; 2]) -> TerrainTile {
     TerrainTile::Static(StaticTerrainTile::Water)
   }
 
-  fn get_tile(&self, world: &World, position: [i32; 2], base_terrain_mod: f64) -> bool {
+  fn get_tile(&self, world: &World, position: [i64; 2], base_terrain_mod: f64) -> bool {
     self.get_value(world, position) - base_terrain_mod > 0.5
   }
 
-  fn get_value(&self, world: &World, [x, y]: [i32; 2]) -> f64 {
+  fn get_value(&self, world: &World, [x, y]: [i64; 2]) -> f64 {
     const NOISE_SCALE: f64 = 0.02;
     world.noise_gen.get([x as f64 * NOISE_SCALE, y as f64 * NOISE_SCALE])
   }
